% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar.R
\name{wkradar_command}
\alias{wkradar_command}
\title{Apply a command JSON and return the resulting response JSON.}
\usage{
wkradar_command(radar, cmd_json)
}
\arguments{
\item{radar}{A radar handle from [wkradar_new()].}

\item{cmd_json}{A command JSON string.}
}
\value{
The response as a JSON string.
}
\description{
Apply a command JSON and return the resulting response JSON.
}
